requirements.txt
mistralai
python-dotenv
pymupdf
pillow
tqdm
